fix setup to have only one testfile per operator including the generation of diagrams
