--- src/operators/debounce-if.test.diagram.ts
@@ -1,18 +1,0 @@
-import { expectObservable, hot } from "../../spec/helpers/marble-testing";
-import { debounceIf } from "./debounce-if";
-import { isEven } from "./debounce-if.test";
-import { TestScheduler } from "rxjs/testing";
-
-declare function asDiagram(arg: string): Function;
-declare const rxTestScheduler: TestScheduler;
-
-describe("diagram", () => {
-    asDiagram("debounceIf(20, n => n % 2 === 0)")("should debounce only even values", () => {
-        // @formatter:off
-        const e1 =   hot("-1--2---3--|");
-        const expected = "-1----2-3--|";
-        // @formatter:on
-
-        expectObservable(e1.pipe(debounceIf(20, isEven, rxTestScheduler))).toBe(expected);
-    });
-});
