+++ src/operators/skip-property-undefined.test.ts
@@ -2,9 +2,24 @@ import { assert } from "chai";
 import { of } from "rxjs";
 import { first, tap } from "rxjs/operators";
 import { skipPropertyUndefined } from "./skip-property-undefined";
+import { expectObservable, hot } from "../../spec/helpers/marble-testing";
 
+declare function asDiagram(arg: string): Function;
 
-describe("skipPropertyNull", function () {
+describe("diagram skipPropertyUndefined", () => {
+    asDiagram(`skipPropertyUndefined(b)`)("should skip values when property b is undefined", () => {
+        const a = { a: 1, b: 2 };
+        const b = { a: 1, b: undefined };
+        // @formatter:off
+        const e1 =   hot("a-b-a|", {a, b});
+        const expected = "a---a|";
+        // @formatter:on
+
+        expectObservable(e1.pipe(skipPropertyUndefined("b"))).toBe(expected, { a });
+    });
+});
+
+describe("skipPropertyUndefined", function () {
 
   it("should should pass object with just values", async function () {
     const testObj = {
