+++ src/operators/skip-some-property-null.test.ts
@@ -2,6 +2,23 @@ import { map } from "rxjs/operators";
 import { assert } from "chai";
 import { of } from "rxjs";
 import { hasNoNullProperties, skipSomePropertyNull } from "./skip-some-property-null";
+import { expectObservable, hot } from "../../spec/helpers/marble-testing";
+
+declare function asDiagram(arg: string): Function;
+
+describe("diagram skipSomePropertyNull", () => {
+  asDiagram(`skipSomePropertyNull()`)("should skip values when any property is null", () => {
+    const a = { a: 1, b: 2 };
+    const b = { a: 1, b: null };
+    const c = { a: null, b: 1 };
+    // @formatter:off
+    const e1 =   hot("a-b-c-a|", {a, b, c});
+    const expected = "a-----a|";
+    // @formatter:on
+
+    expectObservable(e1.pipe(skipSomePropertyNull())).toBe(expected, { a });
+  });
+});
 
 describe("hasNoNullProperties", function () {
